+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+
+- `v0.0.3` Static Icon component
+  - Sizing
+- `v0.0.2` Action component
+  - Responsive behaviour
+
 ## 0.1.0 - 2024-04-16
 ### Added
 - License & copyright